requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.8
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
psycopg2==2.8.6
python-dotenv==0.19.0
gunicorn==20.1.0